Surepetcare, add entity class
